sabba_multi_knowledge_single_model/requirements.txt
python-dotenv==1.0.1
griptape==1.7.3